+++ CHANGES.md
@@ -6,6 +6,7 @@ Key changes in each release. The dates with each release are the dates I submitt
 - minor interface changes
 - improve known word in gist lookup so it now can handle phrases
 - various efficiency improvements
+- flashcards always show current state
 
 *0.0.4 2022-1-2*
 - added forward and back hotkeys for moving about in note history